+++ app/src/app.js
@@ -10,6 +10,7 @@ const ErrorSerializer = require('serializers/error.serializer');
 const sleep = require('sleep');
 const koaValidate = require('koa-validate');
 const koaBody = require('koa-body');
+
 let mongooseOptions = require('../../config/mongoose');
 
 const mongoUri = process.env.MONGO_URI || `mongodb://${config.get('mongodb.host')}:${config.get('mongodb.port')}/${config.get('mongodb.database')}`;