+++ app/test/e2e/dataset-search.spec.js
@@ -2,9 +2,9 @@
 const nock = require('nock');
 const chai = require('chai');
 const Dataset = require('models/dataset.model');
-const { createDataset, deserializeDataset, mapDatasetToMetadataSearchResult } = require('./utils');
+const { createDataset, deserializeDataset, mapDatasetToMetadataSearchResult } = require('./utils/helpers');
 
-const { getTestServer } = require('./test-server');
+const { getTestServer } = require('./utils/test-server');
 
 const should = chai.should();
 
@@ -28,7 +28,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search for common elements in name and description should return 2 results (no synonyms)', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=fake%20dataset`)
             .reply(200, {
                 data: [
@@ -37,7 +37,7 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=fake%20dataset`)
             .reply(200, {
                 data: []
@@ -57,7 +57,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search for name of one dataset should return a single result (no synonyms)', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=${encodeURIComponent(cartoFakeDataset.name)}`)
             .reply(200, {
                 data: [
@@ -65,7 +65,7 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=${encodeURIComponent(cartoFakeDataset.name)}`)
             .reply(200, {
                 data: []
@@ -87,7 +87,7 @@ describe('Search datasets tests', () => {
      * Sort tests
      */
     it('Search with keyword and no explicit sort', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword`)
             .once()
             .reply(200, {
@@ -97,7 +97,7 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=keyword`)
             .once()
             .reply(200, {
@@ -118,7 +118,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search with keyword and sort by most viewed is sorted using the graph', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword`)
             .twice()
             .reply(200, {
@@ -128,14 +128,14 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=keyword&sort=-most-viewed`)
             .twice()
             .reply(200, {
                 data: []
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-datasets-ids?search=keyword&sort=-most-viewed`)
             .once()
             .reply(200, {
@@ -158,7 +158,7 @@ describe('Search datasets tests', () => {
         datasetIdsOne[1].should.equal(cartoFakeDataset._id);
 
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-datasets-ids?search=keyword&sort=-most-viewed`)
             .once()
             .reply(200, {
@@ -182,7 +182,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search with keyword and sort by most favorited is sorted using the graph', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword`)
             .twice()
             .reply(200, {
@@ -192,14 +192,14 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=keyword&sort=-most-favorited`)
             .twice()
             .reply(200, {
                 data: []
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-datasets-ids?search=keyword&sort=-most-favorited`)
             .once()
             .reply(200, {
@@ -222,7 +222,7 @@ describe('Search datasets tests', () => {
         datasetIdsOne[1].should.equal(cartoFakeDataset._id);
 
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-datasets-ids?search=keyword&sort=-most-favorited`)
             .once()
             .reply(200, {
@@ -246,7 +246,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search with keyword and sort by relevance non-specified is sorted using metadata', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword&sort=relevance`)
             .once()
             .reply(200, {
@@ -256,7 +256,7 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=keyword&sort=relevance`)
             .once()
             .reply(200, {
@@ -278,7 +278,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search with keyword and sort by relevance descending is sorted using metadata', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword&sort=-relevance`)
             .once()
             .reply(200, {
@@ -288,7 +288,7 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=keyword&sort=-relevance`)
             .once()
             .reply(200, {
@@ -311,7 +311,7 @@ describe('Search datasets tests', () => {
 
 
     it('Search with keyword and sort by relevance ascending return an error for invalid request', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword&sort=+relevance`)
             .once()
             .reply(400, {
@@ -331,7 +331,7 @@ describe('Search datasets tests', () => {
     });
 
     it('Search with keyword and sort by other keyword is sorted using the metadata', async () => {
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword&sort=metadata`)
             .once()
             .reply(200, {
@@ -341,7 +341,7 @@ describe('Search datasets tests', () => {
                 ]
             });
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/graph/query/search-by-label-synonyms?search=keyword&sort=metadata`)
             .twice()
             .reply(200, {
@@ -360,7 +360,7 @@ describe('Search datasets tests', () => {
         datasetIdsOne[0].should.equal(cartoFakeDataset._id);
         datasetIdsOne[1].should.equal(jsonFakeDataset._id);
 
-        nock(`${process.env.CT_URL}`)
+        nock(process.env.CT_URL)
             .get(`/v1/metadata?search=keyword&sort=metadata`)
             .once()
             .reply(200, {